refactor: move ClassesCancel-related props from ClassCalendar to ClassesCancel
