+++ frontend/src/app/components/customers-dashboard/classes/ClassCalendar.tsx
@@ -1,16 +1,8 @@
 import React, { useEffect, useState, useCallback } from "react";
 import CalendarView from "@/app/components/features/calendarView/CalendarView";
 import styles from "./ClassCalendar.module.scss";
-import {
-  formatFiveMonthsLaterEndOfMonth,
-  isPastPreviousDayDeadline,
-} from "@/app/helper/utils/dateUtils";
-import {
-  cancelClass,
-  fetchClassesForCalendar,
-} from "@/app/helper/api/classesApi";
-import { toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import { formatFiveMonthsLaterEndOfMonth } from "@/app/helper/utils/dateUtils";
+import { fetchClassesForCalendar } from "@/app/helper/api/classesApi";
 import Loading from "../../elements/loading/Loading";
 import ClassActions from "./classActions/ClassActions";
 
@@ -26,12 +18,8 @@ function ClassCalendar({
   const [rebookableClasses, setRebookableClasses] = useState<
     ClassForCalendar[] | undefined
   >(undefined);
-  const [selectedClasses, setSelectedClasses] = useState<
-    { classId: number; classDateTime: string }[]
-  >([]);
   const [isBookableClassesModalOpen, setIsBookableClassesModalOpen] =
     useState(false);
-  const [isCancelingModalOpen, setIsCancelingModalOpen] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
 
@@ -112,70 +100,6 @@ function ClassCalendar({
     fetchData();
   }, [fetchData]);
 
-  const handleCancelingModalClose = () => {
-    setIsCancelingModalOpen(false);
-    setSelectedClasses([]);
-  };
-
-  const handleBulkCancel = async () => {
-    if (selectedClasses.length === 0) return;
-
-    // Get classes that have passed the previous day cancelation deadline
-    const pastPrevDayClasses = selectedClasses.filter((eachClass) =>
-      isPastPreviousDayDeadline(eachClass.classDateTime, "Asia/Tokyo"),
-    );
-
-    if (pastPrevDayClasses.length > 0) {
-      alert(
-        "Classes cannot be canceled on or after the scheduled day of the class.",
-      );
-      const pastPrevDayClassIds = new Set(
-        pastPrevDayClasses.map((pastClass) => pastClass.classId),
-      );
-      const updatedSelectedClasses = selectedClasses.filter(
-        (eachClass) => !pastPrevDayClassIds.has(eachClass.classId),
-      );
-      return setSelectedClasses(updatedSelectedClasses);
-    }
-
-    // Get classes that are before the previous day's deadline
-    const classesToCancel = selectedClasses.filter(
-      (eachClass) =>
-        !isPastPreviousDayDeadline(eachClass.classDateTime, "Asia/Tokyo"),
-    );
-
-    if (classesToCancel.length > 0) {
-      const confirmed = window.confirm(
-        `Are you sure you want to cancel these ${selectedClasses.length} classes?`,
-      );
-      if (!confirmed) return handleCancelingModalClose();
-      try {
-        await Promise.all(
-          classesToCancel.map((eachClass) => cancelClass(eachClass.classId)),
-        );
-        setSelectedClasses([]);
-
-        // Re-fetch data to update the state
-        fetchData();
-        handleCancelingModalClose();
-        toast.success("The classes have been successfully canceled!");
-      } catch (error) {
-        console.error("Failed to cancel classes:", error);
-        setError("Failed to cancel the classes. Please try again later.");
-      }
-    }
-  };
-
-  const toggleSelectClass = (classId: number, classDateTime: string) => {
-    setSelectedClasses((prev) => {
-      const updated = prev.filter((item) => item.classId !== classId);
-      if (updated.length === prev.length) {
-        updated.push({ classId, classDateTime });
-      }
-      return updated;
-    });
-  };
-
   if (error) return <div>{error}</div>;
 
   return (
@@ -185,18 +109,14 @@ function ClassCalendar({
       ) : (
         <>
           <ClassActions
-            setIsCancelingModalOpen={setIsCancelingModalOpen}
             setIsBookableClassesModalOpen={setIsBookableClassesModalOpen}
             rebookableClasses={rebookableClasses}
             isAdminAuthenticated={isAdminAuthenticated}
             customerId={customerId}
             classes={classes}
-            handleBulkCancel={handleBulkCancel}
-            handleCancelingModalClose={handleCancelingModalClose}
-            isCancelingModalOpen={isCancelingModalOpen}
-            selectedClasses={selectedClasses}
-            toggleSelectClass={toggleSelectClass}
             isBookableClassesModalOpen={isBookableClassesModalOpen}
+            fetchData={fetchData}
+            setError={setError}
           />
 
           <CalendarView
