+++ frontend/src/app/components/customers-dashboard/classes/classActions/ClassActions.tsx
@@ -3,33 +3,25 @@ import CancelClasses from "./cancelClasses/CancelClasses";
 import BookClass from "./bookClass/BookClass";
 
 export type ClassActionsProps = {
-  setIsCancelingModalOpen: (open: boolean) => void;
   setIsBookableClassesModalOpen: (open: boolean) => void;
   rebookableClasses?: { id: number; dateTime: string }[];
   isAdminAuthenticated?: boolean;
   customerId: number;
-  isCancelingModalOpen: boolean;
-  handleCancelingModalClose: () => void;
   classes: any;
-  selectedClasses: { classId: number; classDateTime: string }[];
-  toggleSelectClass: (classId: number, classDateTime: string) => void;
-  handleBulkCancel: () => Promise<void>;
   isBookableClassesModalOpen: boolean;
+  fetchData: () => void;
+  setError: (error: string) => void;
 };
 
 export default function ClassActions({
-  setIsCancelingModalOpen,
   setIsBookableClassesModalOpen,
   rebookableClasses,
   isAdminAuthenticated,
   customerId,
-  isCancelingModalOpen,
   isBookableClassesModalOpen,
-  handleCancelingModalClose,
   classes,
-  selectedClasses,
-  toggleSelectClass,
-  handleBulkCancel,
+  fetchData,
+  setError,
 }: ClassActionsProps) {
   return (
     <div className={styles.calendarHeaderContainer}>
@@ -38,12 +30,8 @@ export default function ClassActions({
           <CancelClasses
             classes={classes}
             customerId={customerId}
-            handleBulkCancel={handleBulkCancel}
-            handleCancelingModalClose={handleCancelingModalClose}
-            isCancelingModalOpen={isCancelingModalOpen}
-            selectedClasses={selectedClasses}
-            setIsCancelingModalOpen={setIsCancelingModalOpen}
-            toggleSelectClass={toggleSelectClass}
+            fetchData={fetchData}
+            setError={setError}
           />
 
           <BookClass
